requirements.txt
fastapi
uvicorn
pandas
sqlite3
requests
streamlit
sentence-transformers
faiss-cpu
textblob